move bin into its own mod; move bin style structs into style mod; rename/rework style position field
